Fix useEffect missing dependency warning

+++ src/App.js
@@ -1,32 +1,18 @@
 import React, { useState, useEffect } from 'react';
-import './App.css';
 import { useSearchParams, Outlet } from 'react-router-dom';
+import './App.css';
 import Header from './components/Header';
 import SignIn from './components/Signin';
 import axios from 'axios';
 
 function App() {
   const [user, setUser] = useState(null);
-  const [searchParams,setSearchParms] = useSearchParams();  // eslint-disable-line
-  /* eslint-disable no-debugger, no-console */
+  const [params,setSearchParms] = useSearchParams();
+  setSearchParms(params);
   useEffect(() => {
-    const token = searchParams.get('token');
-    if (token) {
-      axios.get(`${process.env.REACT_APP_APIURL}/login/callback?token=${token}`, {}, {
-        headers: { 'Content-Type': 'application/json'}
-      })
-        .then(res =>{
-            const user = res.data;
-            setUser(user);
-            sessionStorage.setItem('user', JSON.stringify(user));
-            sessionStorage.setItem('token', token);
-        })
-        .catch(error => {
-          sessionStorage.clear();
-        });
-    } else {  // TODO validate sessionStorage token expiration
-      const prevToken = sessionStorage.getItem('token');
-      if(prevToken) {
+    function checkStatus() {
+      const token = params.get('token');
+      if (token) {
         axios.get(`${process.env.REACT_APP_APIURL}/login/callback?token=${token}`, {}, {
           headers: { 'Content-Type': 'application/json'}
         })
@@ -34,15 +20,33 @@ function App() {
               const user = res.data;
               setUser(user);
               sessionStorage.setItem('user', JSON.stringify(user));
-              sessionStorage.setItem('token', prevToken);
+              sessionStorage.setItem('token', token);
           })
           .catch(error => {
-            sessionStorage.clear();  // clean up
+            sessionStorage.clear();
           });
+      } else {  // TODO validate sessionStorage token expiration
+        const prevToken = sessionStorage.getItem('token');
+        if(prevToken) {
+          axios.get(`${process.env.REACT_APP_APIURL}/login/callback?token=${token}`, {}, {
+            headers: { 'Content-Type': 'application/json'}
+          })
+            .then(res =>{
+                const user = res.data;
+                setUser(user);
+                sessionStorage.setItem('user', JSON.stringify(user));
+                sessionStorage.setItem('token', prevToken);
+            })
+            .catch(error => {
+              sessionStorage.clear();  // clean up
+            });
+        }
       }
     }
-  }, []);
-  /* eslint-disable no-debugger, no-console */
+    checkStatus();
+  }, []);   // eslint-disable-line react-hooks/exhaustive-deps
+  // Above es-lint-disable rule is required to avoid useEffect WARNING
+  // that breaks Github Actions
 
   return (
     <div className="container">
